Move bot code to a new package game.ai.bot
