+++ server/src/main/java/game/Game.java
@@ -3,8 +3,8 @@ package game;
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Multimaps;
-import game.ai.Bot;
-import game.ai.BotFactory;
+import game.ai.bot.Bot;
+import game.ai.bot.BotFactory;
 import game.snake.*;
 import game.world.Collidable;
 import game.world.World;
