+++ server/src/main/java/game/ai/bot/Bot.java
@@ -1,4 +1,4 @@
-package game.ai;
+package game.ai.bot;
 
 import game.snake.Snake;
 import game.snake.SnakeChunk;
